admin/requirements.txt
starlette-admin==0.14.0
fastapi==0.115.0
uvicorn==0.30.0
asyncpg==0.30.0
python-dotenv==1.1.0
itsdangerous==2.2.0
python-multipart==0.0.9
httpx==0.27.2
